Adapt huge bid trading

+++ app/src/threads/bidirtrading/bidirtradingthread.h
@@ -67,7 +67,7 @@ public:
     void sellWithPrice(const Quotation& price);
     void buyWithPrice(qint64 amountOfLots, const Quotation& price);
 
-    bool isNeedToSellAsap(qint64 timestamp, BiDirMode mode, float part, float yield, float commission);
+    bool isNeedToSellAsap(qint64 timestamp, BiDirMode mode, double part, double yield, double commission);
 
 private:
     void calculateTotalCostAndInstrumentCost(
@@ -85,7 +85,7 @@ private:
     void calculateBuySellPriceAndLots(
         const tinkoff::GetOrderBookResponse& tinkoffOrderBook,
         const tinkoff::PortfolioResponse&    tinkoffPortfolio,
-        float                                commission,
+        double                               commission,
         qint64&                              lotsToBuy,
         qint64&                              lotsToSell,
         Quotation&                           buyPrice,
@@ -98,7 +98,7 @@ private:
         double                               totalCost,
         double                               instrumentCost,
         double                               instrumentAvgPrice,
-        float                                commission
+        double                               commission
     );
     qint64                calculateLotsToKeep(BiDirMode mode, double totalCost, double bidPrice);
     IDecisionMakerConfig* chooseDecisionConfig();
